объединить презентеры film-details и films, реализовать закрытие попапа, рендеринг комментариев
