+++ src/main.js
@@ -8,10 +8,12 @@ import FooterStatisticsView from './view/footer-statistics-view.js';
 import FilmsPresenter from './presenter/films-presenter.js';
 
 import FilmsModel from './model/films-model.js';
+import CommentsModel from './model/comments-model.js';
 
 const filmsPresenter = new FilmsPresenter();
 
 const filmsModel = new FilmsModel();
+const commentsModel = new CommentsModel(filmsModel);
 
 const siteHeaderElement = document.querySelector('.header');
 const siteMainElement = document.querySelector('.main');
@@ -23,4 +25,4 @@ render(new MainNavigationView(), siteMainElement);
 render(new SortView(), siteMainElement);
 render(new FooterStatisticsView(), footerStatisticsElement);
 
-filmsPresenter.init(siteMainElement, filmsModel);
+filmsPresenter.init(siteMainElement, filmsModel, commentsModel);
