--- src/view/film-details-controls-view.js
@@ -1,28 +1,0 @@
-import { createElement } from '../render.js';
-
-const createFilmDetailsControlsTemplate = () =>
-  `<section class="film-details__controls">
-    <button type="button" class="film-details__control-button film-details__control-button--watchlist" id="watchlist" name="watchlist">Add to watchlist</button>
-    <button type="button" class="film-details__control-button film-details__control-button--active film-details__control-button--watched" id="watched" name="watched">Already watched</button>
-    <button type="button" class="film-details__control-button film-details__control-button--favorite" id="favorite" name="favorite">Add to favorites</button>
-</section>`;
-
-export default class FilmDetailsControlsView {
-  #element = null;
-
-  get template() {
-    return createFilmDetailsControlsTemplate();
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
-  removeElement() {
-    this.#element = null;
-  }
-}
